+++ src/querystring.js
@@ -9,8 +9,7 @@ export function useQueryParams(
 ) {
   const [querystring, setQuerystring] = useState(getQueryString())
   const setQueryParams = useCallback(
-    // TODO: V2 using options param for replace
-    (params, replace = true) => {
+    (params, { replace = true } = {}) => {
       let path = getCurrentPath()
       params = replace ? params : { ...parseFn(querystring), ...params }
       const serialized = serializeFn(params).toString()
@@ -21,25 +20,21 @@ export function useQueryParams(
     [querystring]
   )
   // Update state when route changes
-  useLocationChange(() => setQuerystring(getQueryString()))
+  const updateQuery = useCallback(() => setQuerystring(getQueryString()), [
+    setQueryParams
+  ])
+  useLocationChange(updateQuery)
   return [parseFn(querystring), setQueryParams]
 }
 
 function parseQuery(querystring) {
-  return [...new URLSearchParams(querystring)].reduce(
-    (result, [key, value]) => {
-      result[key] = value
-      return result
-    },
-    {}
-  )
+  return Object.fromEntries(new URLSearchParams(querystring).entries())
 }
 
 function serializeQuery(queryParams) {
-  return Object.entries(queryParams).reduce((query, [key, value]) => {
-    if (value !== null) query.append(key, value)
-    return query
-  }, new URLSearchParams())
+  return new URLSearchParams(
+    Object.entries(queryParams).filter(([, v]) => v !== null)
+  )
 }
 
 export function getQueryString() {
@@ -49,17 +44,4 @@ export function getQueryString() {
     return queryIndex === -1 ? '' : ssrPath.substring(queryIndex + 1)
   }
   return location.search
-}
-
-export function deriveQueryString(current, target, merge) {
-  current = queryAsObject(current)
-  target = queryAsObject(target)
-  return merge
-    ? new URLSearchParams({ ...current, ...target }).toString()
-    : new URLSearchParams(target).toString()
-}
-
-function queryAsObject(query) {
-  if (!query) return {}
-  return Object.fromEntries(new URLSearchParams(query).entries())
 }