+++ test/querystring.spec.js
@@ -1,7 +1,6 @@
 import React from 'react'
 import { render, act, fireEvent } from '@testing-library/react'
 import { navigate, useQueryParams } from '../src/main.js'
-import { deriveQueryString } from '../src/querystring.js'
 
 beforeEach(() => {
   act(() => navigate('/'))
@@ -38,7 +37,10 @@ describe('setQueryParams', () => {
   function Route({ replace, foo = 'bar' }) {
     let [query, setQuery] = useQueryParams()
     return (
-      <button data-testid="update" onClick={() => setQuery({ foo }, replace)}>
+      <button
+        data-testid="update"
+        onClick={() => setQuery({ foo }, { replace })}
+      >
         Set Query: {query.foo}
       </button>
     )
@@ -87,34 +89,5 @@ describe('setQueryParams', () => {
     const { getByTestId } = render(<Route replace={false} />)
     act(() => void fireEvent.click(getByTestId('update')))
     expect(document.location.hash).toEqual('#test')
-  })
-})
-
-describe('deriveQueryString', () => {
-  test('merges current and target when merge is true', () => {
-    expect(deriveQueryString({ a: 'a' }, { b: 'b' }, true)).toBe('a=a&b=b')
-    expect(deriveQueryString('a=a', { b: 'b' }, true)).toBe('a=a&b=b')
-    expect(deriveQueryString({ a: 'a' }, 'b=b', true)).toBe('a=a&b=b')
-    expect(deriveQueryString('a=a', 'b=b', true)).toBe('a=a&b=b')
-    expect(
-      deriveQueryString(
-        new URLSearchParams({ a: 'a' }),
-        new URLSearchParams({ b: 'b' }),
-        true
-      )
-    ).toBe('a=a&b=b')
-  })
-  test('returns target when merge is false', () => {
-    expect(deriveQueryString({ a: 'a' }, { b: 'b' }, false)).toBe('b=b')
-    expect(deriveQueryString('a=a', { b: 'b' }, false)).toBe('b=b')
-    expect(deriveQueryString({ a: 'a' }, 'b=b', false)).toBe('b=b')
-    expect(deriveQueryString('a=a', 'b=b', false)).toBe('b=b')
-    expect(
-      deriveQueryString(
-        new URLSearchParams({ a: 'a' }),
-        new URLSearchParams({ b: 'b' }),
-        false
-      )
-    ).toBe('b=b')
   })
 })